requirements.txt
Flask==3.1.2
pillow==12.0.0
python-dotenv==1.2.1
qrcode==8.2
requests==2.32.5
web3==7.14.0
gunicorn
